requirements.txt
alembic==1.13.3
annotated-types==0.7.0
blinker==1.7.0
cachetools==5.5.0
click==8.1.7
Flask==3.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.0.1
grpcio==1.68.0
grpcio-status==1.68.0
httplib2==0.22.0
itsdangerous==2.1.2
Mako==1.3.5
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary>=2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
rsa==4.9
SQLAlchemy==2.0.23
tqdm==4.67.0
uritemplate==4.1.1
Werkzeug==3.0.1
